Added Fetch.js
Fetch is located in the src/Backend folder. It currently contains the User and Portfolio methods used. 

The GET methods return a JSON-parsed object, while the other methods return the JSON response/error objects as they are. This is done for flexibility in handling the responses/errors by attaching a then-catch block at the end of the calls to the non-GET methods.