Has authentication*, refactored Endpoints and Fetch
*Except for Signup.js

Also added Traditional Chinese (Google Translate-d from Simplified Chinese though, so probably needs reviewing)

Still has a few problems:

- Used in PublicPortfolio, the getPublicUser and getPortfolio functions in Fetch.js that use GET /users/{username} and GET /portfolios/{portfolio_id} respectively still needs Bearer access token to run, but it actually shouldn't:

This is because non-logged-in people must be able to view the public portfolios, which are retrieved using these GETs.

- Need to figure out how to not just add a route on window.location.href when logging in. (Server-side rendering?):

This is because every time we refresh when we are logged in (i.e. sessionStorage.getItem('accessToken') returns a proper accessToken string that is still valid and can get a proper user when we run '/getUser'), the page is still on the '/login' route and and we are just being redirected to '/portfolios' because we are now loggedIn. This isn't the rendering that we want.