Themes for public portfolios using PublicThemesContext
Currently saves the chosen theme to localStorage.getItem('publicTheme'), only 3 themes available as of now: Light, Dark, and Muted Light

Will need to refactor so that the portfolio is patched if another theme is chosen, saving the new theme in a 'theme' property in each Portfolio object (not implemented yet in backend).

Hierarchy where PublicThemesProvider (such that its children components can access and modify the public theme of the current portfolio) is used:
- PublicThemesProvider > CssBaseline > PublicPortfolio
- PublicThemesProvider > ThemesProvider > CssBaseline > EditPortfolioStyle

In all cases, a component will use the colour provided by the closest/most immediate theme provider, e.g. EditPortfolioStyle will be rendered using the colours of the theme in ThemesProvider, while PublicPortfolio will be rendered using that of PublicThemesProvider.